Fix: resolve all -Wsign-conversion 'changes value' sentinels (Bucket 6)
Hand-audited each of the 13 sign-flip-to-all-ones warnings; all are
intentional sentinels or type mismatches, none are missing error checks:

dkim-test.c -- dkim_test_dns_get() returns size_t but used bare -1 as
the 'nothing queued' sentinel.  The sole caller (dkim-keys.c) already
compares against (size_t)-1 with a comment explaining the idiom.
Change all nine return sites to return (size_t)-1 to make the intent
explicit and match the call-site convention.

dkim.c:4067 -- sig->sig_flags &= ~DKIM_SIGFLAG_PASSED: standard
bit-clear on a u_int field; ~(int 0x04) = int -5 promoting to u_int.
Cast the flag to u_int before inverting: ~(u_int)DKIM_SIGFLAG_PASSED.

opendkim-db.c:507 -- db_flags &= ~DKIMF_DB_FLAG_MAKELOCK: same
bit-clear pattern (flag 0x20, ~0x20 = -33 as int).
Fix: ~(u_int)DKIMF_DB_FLAG_MAKELOCK.

opendkim-db.h:28 + opendkim-db.c -- DKIMF_DB_TYPE_UNKNOWN is (-1)
but db_type was u_int; dkimf_db_type() already returns int and the
type-code table uses int.  Change db_type from u_int to int in the
private struct; remove the (u_int) workaround cast from the comparison
at db.c:491.

Found via -Wsign-conversion under strict build
(-DOPENDKIM_ENABLE_STRICT_C=ON -DOPENDKIM_ENABLE_EXTRA_WARNINGS=ON).

+++ libopendkim/dkim-test.c
@@ -155,7 +155,7 @@ dkim_test_dns_get(DKIM *dkim, u_char *buf, size_t buflen)
 
 	td = dkim->dkim_dnstesth;
 	if (td == NULL)
-		return -1;
+		return (size_t)-1;
 
 	dkim->dkim_dnstesth = td->dns_next;
 
@@ -175,7 +175,7 @@ dkim_test_dns_get(DKIM *dkim, u_char *buf, size_t buflen)
 	if (n < 0)
 	{
 		DKIM_FREE(dkim, td);
-		return -1;
+		return (size_t)-1;
 	}
 	cp += n;
 	/*
@@ -186,7 +186,7 @@ dkim_test_dns_get(DKIM *dkim, u_char *buf, size_t buflen)
 	if (cp >= end || (size_t) (end - cp) < 2 * sizeof(uint16_t))
 	{
 		DKIM_FREE(dkim, td);
-		return -1;
+		return (size_t)-1;
 	}
 
 	DKIM_PUTSHORT(td->dns_type, cp);
@@ -207,7 +207,7 @@ dkim_test_dns_get(DKIM *dkim, u_char *buf, size_t buflen)
 	if (n < 0)
 	{
 		DKIM_FREE(dkim, td);
-		return -1;
+		return (size_t)-1;
 	}
 	cp += n;
 
@@ -216,7 +216,7 @@ dkim_test_dns_get(DKIM *dkim, u_char *buf, size_t buflen)
 	    (size_t) (end - cp) < 2 * sizeof(uint16_t) + sizeof(uint32_t))
 	{
 		DKIM_FREE(dkim, td);
-		return -1;
+		return (size_t)-1;
 	}
 
 	DKIM_PUTSHORT(td->dns_type, cp);
@@ -236,7 +236,7 @@ dkim_test_dns_get(DKIM *dkim, u_char *buf, size_t buflen)
 		    (size_t) (end - cp) < (size_t) len + sizeof(uint16_t))
 		{
 			DKIM_FREE(dkim, td);
-			return -1;
+			return (size_t)-1;
 		}
 		DKIM_PUTSHORT(len, cp);
 
@@ -261,14 +261,14 @@ dkim_test_dns_get(DKIM *dkim, u_char *buf, size_t buflen)
 		if (cp >= end || (size_t) (end - cp) < sizeof(uint16_t))
 		{
 			DKIM_FREE(dkim, td);
-			return -1;
+			return (size_t)-1;
 		}
 		DKIM_PUTSHORT(td->dns_prec, cp);
 		n = dn_comp((char *) td->dns_reply, cp, end - cp, NULL, NULL);
 		if (n < 0)
 		{
 			DKIM_FREE(dkim, td);
-			return -1;
+			return (size_t)-1;
 		}
 
 		cp += n;
@@ -276,7 +276,7 @@ dkim_test_dns_get(DKIM *dkim, u_char *buf, size_t buflen)
 
 	  default:
 		DKIM_FREE(dkim, td);
-		return -1;
+		return (size_t)-1;
 	}
 
 	DKIM_FREE(dkim, td);
